@@ -1,9 +1,5 @@
 package se.yawnmedia.fairsplit;
 
-/**
- * Created by Jeppe on 2018-03-06.
- */
-
 public final class Settings {
-    public static String BaseURL = "https://fairsplit.jespergustin.se:54321";
+    public static String BaseURL = "https://fairsplit.minegenserver.se:54321";
 }
